requirements.txt
streamlit
tensorflow-cpu==2.12.0
numpy
pandas
scikit-learn
Pillow
plotly
streamlit-option-menu
matplotlib
seaborn